backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
requests
google-generativeai
pymupdf
python-multipart
